Make the calculator work
Add event listeners to all the number buttons & operators, and
initialize a chosenNumbers array and chosenOperator variable. Clicking
equal's button will call the operate() function with the two numbers
that will be stored in chosenNumbers, and the operator in chosenOperator

Another way to operate is by clicking an operator twice, this will
perform the operation then add in the result to the chosenNumbers array.
This is so that the user can keep chaining operators and the calculator
will do the operations.